Separate CorrelationsTwoPoint into two classes.
Moved catalog loading and cutting functions into their own functions.
Created a plotting Mixin class for the correlation classes.
Split CorrelationsTwoPoint into to classes to avoid if, else cases on
when using redshifts for projected two point correlations. This
simplifies the classes significantly and allows them to be more
flexable in the cuts applied and the comparisons that are made.

Fix typo.

Debug CorrelationsProjectedTwoPoint

Fixed typos.
Fixed data types and inputs.

Debug CorrelationsProjectedTwoPoint.

Compared to pervious version of the test and produced similar
results to within the scatter of the produced randoms.